# Optimize WHITESPACEP, as it is often used.